+++ mmodel/handler.py
@@ -2,6 +2,8 @@ from collections import UserDict
 from mmodel.utility import model_signature, graph_topological_sort, param_counter
 from datetime import datetime
 import h5py
+import string
+import random
 
 ERROR_FORMAT = """\
 Exception occurred for node '{node}':
@@ -51,13 +53,12 @@ class TopologicalHandler:
             else:
                 data.update(dict(zip(returns, func_output)))
         except:
-            # format the error message
-
             try:  # if the data class need to be closed
                 data.close()
             except:
                 pass
 
+            # format the error message
             node_str = self.graph.view_node(node)
             input_str = "\n".join(
                 [f"  {key} = {repr(value)}" for key, value in kwargs.items()]
@@ -82,10 +83,7 @@ class TopologicalHandler:
 
 
 class MemData(UserDict):
-    """Modify dictionary that checks the counter every time a value is accessed
-
-    The iteration method is
-    """
+    """Modify dictionary that checks the counter every time a value is accessed"""
 
     def __init__(self, data, counter):
         """Counter is a copy of the counter dictionary"""
@@ -110,20 +108,13 @@ class MemData(UserDict):
         else:
             return super().__getitem__(key)
 
-    # def __iter__(self):
-    #     """Prevent iteration that reduces the counter"""
-    #     raise RuntimeError("MemDict is not iterable")
-
 
 class H5Data:
     """Data class to interact with underlying h5 file
 
-    :param str fname: h5 file name
-    :param str gname: group name for the data entry
-
-    The timestamp is appended to gname to ensure unique entry
-    (in the case that the function runs really fast, a 0 is appended
-    to the end of the file name.)
+    The timestamp-uuid is used to ensure unique entries to group.
+    The randomly generated short uuid has 36^5, which is roughly 2e9
+    possibilities (picoseconds range).
     """
 
     def __init__(self, data, fname, gname):
@@ -131,14 +122,15 @@ class H5Data:
         self.fname = fname
 
         self.f = h5py.File(self.fname, "a")
-        try:
-            self.gname = f"{gname} {datetime.now()}"
-            self.group = self.f.create_group(self.gname)
-        except:
-            # add a digit to time (might not be the best solution)
-            self.gname += "0"
-            self.group = self.f.create_group(self.gname)
 
+        alphabet = string.ascii_lowercase + string.digits
+        shortuuid = "".join(random.choices(alphabet, k=6))
+        self.gname = f"{gname} {datetime.now().strftime('%y%m%d-%H%M%S')}-{shortuuid}"
+        self.group = self.f.create_group(self.gname)
+        self.update(data)
+
+    def update(self, data):
+        """Write key values in bulk"""
         for key, value in data.items():
             self[key] = value
 
@@ -161,10 +153,6 @@ class H5Data:
         """
         try:
             self.group.create_dataset(key, data=value)
-        # except ValueError:
-        #     # ValueError: Unable to create dataset (name already exists)
-        #     # The error should not occur since all key values are unique
-        #     self.group[key] = value
         except TypeError:
             # TypeError: Object dtype dtype('O') has no native HDF5 equivalent
             self.group.attrs[key] = str(value)
@@ -200,8 +188,8 @@ class MemHandler(TopologicalHandler):
 class H5Handler(TopologicalHandler):
     """H5 Handler, saves all calculation values to a h5 file
 
-    Each entry of the h5 file is unique, a timestamp is added after the
-    given "gname".
+    :param str fname: h5 file name
+    :param str gname: group name for the data entry
     """
 
     DataClass = H5Data